Modificadas las vistas en users
